+++ src/bmk/adapters/cli/commands/__init__.py
@@ -33,6 +33,7 @@ from .cov_cmd import cli_codecov, cli_cov, cli_coverage
 from .custom_cmd import cli_custom
 from .dependencies_cmd import cli_d, cli_dependencies, cli_deps
 from .email import cli_send_email, cli_send_notification
+from .ensure_cmd import cli_ensure
 from .info import cli_fail, cli_info
 from .install_cmd import cli_install
 from .logging import cli_logdemo
@@ -64,6 +65,7 @@ __all__ = [
     "cli_d",
     "cli_dependencies",
     "cli_deps",
+    "cli_ensure",
     "cli_fail",
     "cli_info",
     "cli_install",